+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - wex::quoted uses std::quoted and includes a delim character
 
+### Changed
+
+### Fixed
+
+- macro playback and undo includes sub commands
+
 ## 24.10 - 2024-09-22
 
 ### Added